Hearo/sound_pipeline/requirements.txt
# Sound Pipeline Requirements
# Core dependencies
numpy>=1.21.0
torch>=1.9.0
torchaudio>=0.9.0
transformers>=4.20.0
requests>=2.28.0

# Audio processing
pyaudio>=0.2.11
wave

# USB device control
pyusb>=1.2.1

# Optional: For better performance
# torch-audiomentations>=0.11.0
# librosa>=0.9.0

# Development dependencies (optional)
# pytest>=7.0.0
# matplotlib>=3.5.0
# tqdm>=4.64.0
